… .vscode/settings.json
… Cargo.toml
… examples/all_files.rs
… examples/bench.rs
… examples/example.rs
… examples/specific_file.rs
… src/matching.rs
… src/test.rs
… tests/code.desktop
… tests/com.brave.Browser.flextop.brave-cinhimbnkkaeohfgghhklpknlkffjgod-Default.desktop
… tests/org.mozilla.firefox.desktop