## Feature

- Change the API a bit, `&'a str` to `Cow<'a, str>`, for storing the Entry struct in app state.
- allow passing an array of locales to filter the final entry

In my test, the borrowed api performance are not impacted.

#### Details benchmark

The format is number of desktop file parsed:time. All test are using release mode.

- current main: `332: 7.73ms` (note that this is with the locale filtering optimization, the real result will be around 9)

this PR:
```
bench_borrowed 332: 7.15ms
bench_owned 332: 12.53ms
bench_owned_optimized 332: 7.31ms
```


## Todo

- [x] improve the matching function to take into account more desktop field, and their localization
- [ ] ~~improve the localization api (should not return a `gettext` value behind your back).~~ matching is not impacted
- [ ] ~~feature gate for matching~~
- [ ]  ~~maybe integrate rayon for matching, since it will be quite expensive in the end (need benchmark)~~ already fast for now
- [x] integrate this change in cosmic-launcher to see if it fit well