@mmstick 2024-06-03
I don't think a feature flag is necessary. Those who don't want to use it can simply not use the matching module.