@wiiznokes 2024-06-03
The matching function could be improved once launcher support giving score to the service. This will be easier to test things